+++ docs/source/conf.py
@@ -21,9 +21,9 @@ autoapi_prepare_jinja_env = prepare_jinja_env
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
 project = 'em_map_utils'
-copyright = '2025, Ardan Patwardhan'
+copyright = '2025-2026, Ardan Patwardhan'
 author = 'Ardan Patwardhan'
-release = '0.0.2'
+release = '0.1.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration